requirements.txt
pyyaml
tensorflow-addons
tensorflow-io
